feat: implement minimal API endpoints for the products service + downgrade datalayer to .net 9 as Pomelo does not yet support .net 10 (required for MySQL EfCore integration)

+++ BusinessLogicLayer/Services/ProductsService.cs
@@ -1,12 +1,12 @@
 ﻿using AutoMapper;
 using BusinessLogicLayer.DTO;
 using BusinessLogicLayer.ServiceContracts;
+using BusinessLogicLayer.Validators;
 using DataAccessLayer.Entities;
 using DataAccessLayer.RepositoryContracts;
-using System;
-using System.Collections.Generic;
+using FluentValidation;
+using FluentValidation.Results;
 using System.Linq.Expressions;
-using System.Text;
 
 namespace BusinessLogicLayer.Services
 {
@@ -15,15 +15,43 @@ namespace BusinessLogicLayer.Services
         //dependencies
         private readonly IMapper _mapper;
         private readonly IProductsRepository _productsRepository;
+        private readonly ProductAddRequestValidator _productAddRequestValidator;
+        private readonly ProductUpdateRequestValidator _productUpdateRequestValidator;
 
-        public ProductsService(IMapper mapper, IProductsRepository productsRepository)
+        /// <summary>
+        /// The constructor for the ProductsService class, which initializes the dependencies for the service.
+        /// </summary>
+        /// <param name="mapper">Dependency Injected IMapper instance</param>
+        /// <param name="productsRepository">Dependency Injected IProductsRepository instance</param>
+        /// <param name="productAddRequestValidator">Dependency Injected ProductAddRequestValidator instance - as we use Minimal API, not Controller based API, so we need to manually validate requests</param>
+        /// <param name="productUpdateRequestValidator">Dependency Injected ProductUpdateRequestValidator instance - as we use Minimal API, not Controller based API, so we need to manually validate requests</param>
+        public ProductsService(IMapper mapper, IProductsRepository productsRepository, ProductAddRequestValidator productAddRequestValidator, ProductUpdateRequestValidator productUpdateRequestValidator)
         {
             _mapper = mapper;
             _productsRepository = productsRepository;
+            _productAddRequestValidator = productAddRequestValidator;
+            _productUpdateRequestValidator = productUpdateRequestValidator;
         }
 
         public async Task<ProductResponse?> AddProduct(ProductAddRequest productAddRequest)
         {
+            #region validate the request
+            if (productAddRequest == null)
+            {
+                throw new ArgumentNullException(nameof(productAddRequest));
+            }
+
+            //Validate the product using Fluent Validation
+            ValidationResult validationResult = await _productAddRequestValidator.ValidateAsync(productAddRequest);
+
+            // Check the validation result
+            if (!validationResult.IsValid)
+            {
+                string errors = string.Join(", ", validationResult.Errors.Select(temp => temp.ErrorMessage));
+                throw new ArgumentException(errors);
+            }
+            #endregion
+
             var product = _mapper.Map<Product>(productAddRequest);
             var addedProduct = await _productsRepository.AddProduct(product);
             var productResponse = _mapper.Map<ProductResponse>(addedProduct);
@@ -32,6 +60,12 @@ namespace BusinessLogicLayer.Services
 
         public async Task<bool> DeleteProduct(Guid productID)
         {
+            #region validate the request
+            if (productID == Guid.Empty || 
+                await _productsRepository.GetProductByCondition(p => p.ProductID == productID) == null) 
+                throw new ArgumentException($"Product with ID {productID} does not exist.");
+            #endregion
+
             return await _productsRepository.DeleteProduct(productID);
         }
 
@@ -61,6 +95,24 @@ namespace BusinessLogicLayer.Services
 
         public async Task<ProductResponse?> UpdateProduct(ProductUpdateRequest productUpdateRequest)
         {
+            #region validate the request
+            if (productUpdateRequest == null)
+            {
+                throw new ArgumentNullException(nameof(productUpdateRequest));
+            }
+
+            //Validate the product using Fluent Validation
+            ValidationResult validationResult = await _productUpdateRequestValidator.ValidateAsync(productUpdateRequest);
+
+            // Check the validation result
+            if (!validationResult.IsValid)
+            {
+                string errors = string.Join(", ", validationResult.Errors.Select(temp => temp.ErrorMessage));
+                throw new ArgumentException(errors);
+            }
+            #endregion
+
+
             var product = _mapper.Map<Product>(productUpdateRequest);
             var updatedProduct = await _productsRepository.UpdateProduct(product);
             var productResponse = _mapper.Map<ProductResponse>(updatedProduct);
